\name{Tutorial4-package}
\alias{Tutorial4-package}
\alias{Tutorial4}
\docType{package}
\title{
\packageTitle{Tutorial4}
}
\description{
\packageDescription{Tutorial4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Tutorial4}
\packageIndices{Tutorial4}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Tutorial4}

Maintainer: \packageMaintainer{Tutorial4}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
